training sessions controller has been updated

+++ app/Http/Controllers/TrainingSessionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use yajra\Datatables\Datatables;
 use App\Models\TrainingSession;
+use App\Models\AttendedSession;
 use App\Models\Gym;
 
 
@@ -45,22 +46,31 @@ class TrainingSessionController extends Controller
 
     public function update(Request $request, $id){
         $request_out=$request->all();
-        TrainingSession::where('id', $id)->update([
-            'name'=> $request_out['name'],
-            'gym_id'=> $request_out['gym_id'],
-            'coach_id'=> $request_out['coach_id'],
-            'starts_at'=> $request_out['starts_at'],
-            'finishes_at'=> $request_out['finishes_at'],
-        ]);
+        if(AttendedSession::where("id",$id)->first()){
+            return redirect()->back() ->with('alert', 'This session already had Trainees');
+        }else{
+            TrainingSession::where('id', $id)->update([
+                'name'=> $request_out['name'],
+                'gym_id'=> $request_out['gym_id'],
+                'coach_id'=> $request_out['coach_id'],
+                'starts_at'=> $request_out['starts_at'],
+                'finishes_at'=> $request_out['finishes_at'],
+            ]);
+            return redirect('training_sessions');
+        }
 
-        return redirect('training_sessions');
+        
     }
 
     public function destroy(Request $request)
     {
         $request_out=$request->all();
-        TrainingSession::where("id", $request_out['id'])->delete();
-        return back();
+        if(AttendedSession::where("id",$request_out['id'])->first()){
+            return("maynf34 ya ngm");
+        }else{
+            TrainingSession::where("id", $request_out['id'])->delete();
+            return("removed");
+        }
     }
 
     public function getTrainingSessions()
